refactor(pos-catalog, pos-customer): remove entity types from service interface signatures (#1550)
pos-catalog: ProductLifecycleState and ProductStatus are plain domain
enums that lived under internal.entity only by colocation; move them to
internal.enums and update all references, clearing both service
interfaces of entity-package imports.

pos-customer: CrmVehicleService and PartyService declared
CommercialParty-typed methods that were only ever self-called by their
impls; remove them from the interfaces and demote to package-private
impl methods.
