+++ pos-catalog/src/main/java/com/positivity/catalog/internal/dto/ProductStatusUpdateRequestDto.java
@@ -1,6 +1,6 @@
 package com.positivity.catalog.internal.dto;
 
-import com.positivity.catalog.internal.entity.ProductStatus;
+import com.positivity.catalog.internal.enums.ProductStatus;
 import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 import lombok.Setter;
