+++ pos-catalog/src/main/java/com/positivity/catalog/internal/enums/ProductStatus.java
@@ -1,4 +1,4 @@
-package com.positivity.catalog.internal.entity;
+package com.positivity.catalog.internal.enums;
 
 public enum ProductStatus {
     ACTIVE,
