+++ pos-catalog/src/main/java/com/positivity/catalog/internal/service/ProductLifecycleServiceImpl.java
@@ -5,8 +5,8 @@ import com.positivity.catalog.internal.dto.ProductLifecycleResponse;
 import com.positivity.catalog.internal.dto.ProductLifecycleUpdateRequest;
 import com.positivity.catalog.internal.dto.ProductReplacementRequest;
 import com.positivity.catalog.internal.entity.ProductEntity;
-import com.positivity.catalog.internal.entity.ProductLifecycleState;
 import com.positivity.catalog.internal.entity.ProductReplacementEntity;
+import com.positivity.catalog.internal.enums.ProductLifecycleState;
 import com.positivity.catalog.internal.exception.CatalogBusinessRuleException;
 import com.positivity.catalog.internal.exception.CatalogForbiddenOperationException;
 import com.positivity.catalog.internal.exception.CatalogNotFoundException;
