+++ pos-catalog/src/main/java/com/positivity/catalog/internal/service/ProductMasterDataServiceImpl.java
@@ -11,8 +11,8 @@ import com.positivity.catalog.internal.dto.SubcategoryDto;
 import com.positivity.catalog.internal.entity.Category;
 import com.positivity.catalog.internal.entity.ProductCodeType;
 import com.positivity.catalog.internal.entity.ProductEntity;
-import com.positivity.catalog.internal.entity.ProductStatus;
 import com.positivity.catalog.internal.entity.Subcategory;
+import com.positivity.catalog.internal.enums.ProductStatus;
 import com.positivity.catalog.internal.exception.CatalogBusinessRuleException;
 import com.positivity.catalog.internal.exception.CatalogNotFoundException;
 import com.positivity.catalog.internal.exception.CatalogValidationException;
