+++ pos-catalog/src/test/java/com/positivity/catalog/internal/config/CatalogFactPublisherTest.java
@@ -9,7 +9,6 @@ import static org.mockito.Mockito.when;
 
 import com.positivity.catalog.internal.entity.Category;
 import com.positivity.catalog.internal.entity.ProductEntity;
-import com.positivity.catalog.internal.entity.ProductStatus;
 import com.positivity.catalog.internal.entity.ProductTrackingLevel;
 import com.positivity.catalog.internal.entity.ProductUomEntity;
 import com.positivity.catalog.internal.entity.ProductUomType;
@@ -18,6 +17,7 @@ import com.positivity.catalog.internal.entity.Subcategory;
 import com.positivity.catalog.internal.entity.SubstitutionGroupEntity;
 import com.positivity.catalog.internal.entity.SubstitutionGroupMemberEntity;
 import com.positivity.catalog.internal.entity.SupplierArticleCodeEntity;
+import com.positivity.catalog.internal.enums.ProductStatus;
 import com.positivity.catalog.internal.repository.ProductUomRepository;
 import com.positivity.catalog.internal.repository.SubstitutionGroupMemberRepository;
 import com.positivity.domainevents.DomainEventEnvelope;
